% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PICBcombine.R
\name{PICBcombine}
\alias{PICBcombine}
\title{Combine Granges from few replicate into a set of repeatable intervals}
\usage{
PICBcombine(
  rangesList = NULL,
  alignmentsList = NULL,
  REFERENCE.GENOME = NULL,
  TYPE.OF.REGION = allalignments,
  SEQ.LEVELS.STYLE = "UCSC"
)
}
\arguments{
\item{rangesList}{list of few PICBbuild outputs}

\item{alignmentsList}{named list of PICBload outputs. Names correspond to replicate names.}

\item{REFERENCE.GENOME}{name of genome. For example "BSgenome.Dmelanogaster.UCSC.dm6"}

\item{TYPE.OF.REGION}{"clusters" by default}

\item{SEQ.LEVELS.STYLE}{style of chromosome names for BSgenome. "UCSC" by default.}
}
\value{
Granges object annotated for every replicate.
}
\description{
Combine Granges from few replicate into a set of repeatable intervals
}
\author{
Aleksandr Friman
}
