% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PICBexporttoexcel.R
\name{PICBexporttoexcel}
\alias{PICBexporttoexcel}
\title{Export clustering object into an Excel file}
\usage{
PICBexporttoexcel(RANGES = NULL, EXCEL.FILE.NAME = NULL)
}
\arguments{
\item{RANGES}{clustering object to export}

\item{EXCEL.FILE.NAME}{file name to save}
}
\value{
no values returned
}
\description{
Export clustering object into an Excel file
}
\examples{
PICBexporttoexcel(myclusteringObject, "~/piRNAclustersFromUFOsamples.Area51.xlsx")
}
\author{
Aleksandr Friman
}
