% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PICBimportfromexcel.R
\name{PICBimportfromexcel}
\alias{PICBimportfromexcel}
\title{Import cluster object from an Excel file}
\usage{
PICBimportfromexcel(EXCEL.FILE.NAME = NULL)
}
\arguments{
\item{EXCEL.FILE.NAME}{file name to import from}
}
\value{
list of annotated GRanges objects named "seeds" for seeds,
"cores" for cores,
"clusters" for clusters
}
\description{
Import cluster object from an Excel file
}
\examples{
importedClusters <- PICBimportfromexcel(
    EXCEL.FILE.NAME = system.file("extdata", "myClusters_demonstration.xlsx", package = "PICB")
)
}
\author{
Aleksandr Friman
}
