% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PICBoptimize.R
\name{PICBoptimize}
\alias{PICBoptimize}
\title{Runs PICBbuild multiple times with provided parameters and returns optimization data frame.}
\usage{
PICBoptimize(
  IN.ALIGNMENTS,
  REFERENCE.GENOME,
  VERBOSITY = 2,
  SEQ.LEVELS.STYLE = "UCSC",
  ...
)
}
\arguments{
\item{IN.ALIGNMENTS}{list of alignments from PICBload}

\item{REFERENCE.GENOME}{name of genome. For example "BSgenome.Dmelanogaster.UCSC.dm6"}

\item{VERBOSITY}{verbosity level 0/1/2/3. 2 by default.}

\item{...}{rest of the parameters used by PICBbuild and provided as iteratable vectors}
}
\value{
Optimization values dataframe
}
\description{
Runs PICBbuild multiple times with provided parameters and returns optimization data frame.
}
\examples{
PICBoptimize(IN.ALIGNMENTS = myAlignmentsFromPIBCload, REFERENCE.GENOME="BSgenome.Dmelanogaster.UCSC.dm6", MIN.UNIQUE.ALIGNMENTS.PER.WINDOW=c(1,2,3,4,5))
}
\author{
Aleksandr Friman
}
