% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PICBgetchromosomes.R
\name{PICBgetchromosomes}
\alias{PICBgetchromosomes}
\title{Get SeqInfo object from BSgenome name}
\usage{
PICBgetchromosomes(BS.SPECIES, SEQ.LEVELS.STYLE = "UCSC")
}
\arguments{
\item{BS.SPECIES}{name of genome. For example "BSgenome.Dmelanogaster.UCSC.dm6"}

\item{SEQ.LEVELS.STYLE}{naming of chromosomes style. "UCSC" by default}
}
\value{
SeqInfo object with standard non-circular chromosome names
}
\description{
Get SeqInfo object from BSgenome name
}
\examples{
library(BSgenome.Dmelanogaster.UCSC.dm6)
mySI<-PICBgetchromosomes("BSgenome.Dmelanogaster.UCSC.dm6", "UCSC")
}
\author{
Aleksandr Friman
}
