% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PICBimportfromexcel.R
\name{PICBimportfromexcel}
\alias{PICBimportfromexcel}
\title{Import clustering object from an Excel file}
\usage{
PICBimportfromexcel(EXCEL.FILE.NAME = NULL)
}
\arguments{
\item{EXCEL.FILE.NAME}{file name to import from}
}
\value{
list of annotated Granges objects named "seeds" for seeds,
"cores" for cores,
"clusters" for clusters
}
\description{
Import clustering object from an Excel file
}
\author{
Aleksandr Friman
}
