% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PICBloadfasta.R
\name{PICBloadfasta}
\alias{PICBloadfasta}
\title{Get SeqInfo object from a fasta file}
\usage{
PICBloadfasta(FASTA.NAME = NULL)
}
\arguments{
\item{FASTA.NAME}{path to the fasta file}
}
\value{
SeqInfo object with all chromosome names and lengths from the fasta file
}
\description{
Get SeqInfo object from a fasta file
}
\examples{
mySI<-PICBloadfasta("~/path/to/your.fasta")
}
