% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PICBstrandanalysis.R
\name{PICBstrandanalysis}
\alias{PICBstrandanalysis}
\title{Add sense/antisense ratio of piRNAs per piRNAcluster}
\usage{
PICBstrandanalysis(IN.ALIGNMENTS, IN.RANGES, VERBOSE = TRUE)
}
\arguments{
\item{IN.ALIGNMENTS}{list of alignments from PICBload}

\item{IN.RANGES}{single GRanges object (seeds, cores or clusters from PICBbuild)}

\item{VERBOSE}{enables progress output. True by default.}
}
\value{
Granges object with an additional annotation column
}
\description{
Add sense/antisense ratio of piRNAs per piRNAcluster
}
\examples{
outputOfPICBbuild$clusters<-PICBstrandanalysis(outputOfPICBload, outputOfPICBbuild$clusters)
}
