% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PICBstrandanalysis.R
\name{PICBstrandanalysis}
\alias{PICBstrandanalysis}
\title{Indicate Granges dual stranded clusters and bi-directional clusters}
\usage{
PICBstrandanalysis(
  INPUT.GRANGES,
  DUAL.STRAND.MIN.OVERLAP = 1,
  PROD.FOLD = 3,
  BIDECTIONAL.MAX.OVERLAP = 2000,
  TSS.LEN = 1000
)
}
\arguments{
\item{INPUT.GRANGES}{Granges (seeds/cores/clusters) to annotate}

\item{DUAL.STRAND.MIN.OVERLAP}{min overlap for dual stranded clusters}

\item{PROD.FOLD}{max difference in productivity (all 1ry alignments FPKM) for dual-strand clusters}

\item{BIDECTIONAL.MAX.OVERLAP}{max overlap for bi-directional clusters}

\item{TSS.LEN}{length of promoter}
}
\value{
Granges object with extra annotation columns
}
\description{
Indicate Granges dual stranded clusters and bi-directional clusters
}
\examples{
outputOfPICBbuild$clusters<-PICBstrandanalysis(outputOfPICBbuild$clusters)
}
\author{
Aleksandr Friman

Parthena Konstantinidou

Zuzana Loubalova
}
